% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\docType{data}
\name{bacteria}
\alias{bacteria}
\title{Raw Taxon Count Table}
\format{
An object of class \code{data.frame} with 63 rows and 139 columns.
}
\usage{
bacteria
}
\description{
The data is a raw taxon count table consisting of 63 samples and 139 taxa.
}
\keyword{datasets}
