% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_selection.R
\name{balance_selection}
\alias{balance_selection}
\title{Selection of the optimal balance of microbiome associated with survival outcomes}
\usage{
balance_selection(
  Surv_obj,
  data,
  covariates = NULL,
  min_prevalence = 0.1,
  mult_repl = TRUE,
  model = "coxph",
  dist = "weibull",
  stopping_pvalue = 0.15,
  sequential_test = FALSE,
  sequential_alpha = 0.25,
  selection_criterion = "min_decrement_pvalue",
  selection_threshold = 0.15,
  quantile_plotted = c(0.25, 0.75),
  alpha = 0.05
)
}
\arguments{
\item{Surv_obj}{An object of class \code{Surv} generated by
\code{\link[survival:Surv]{survival::Surv}}}

\item{data}{A \eqn{n \times p} matrix which is the raw taxon count table,
where rows are the samples, columns are the taxa}

\item{covariates}{A \eqn{n \times q} data frame containing important covariates for adjustment,
which could be a mixture of \eqn{q} continuous and discrete (pre-defined as factor) variables.
The default value is \code{NULL}}

\item{min_prevalence}{The minimum prevalence of taxa that could be considered
in the selection procedure. The default value is \code{0.1}}

\item{mult_repl}{A logical variable indicating which algorithm is used to
process the raw taxon count table. If \code{TRUE} then the geometric Bayesian multiplicative
replacement (\code{\link[zCompositions:cmultRepl]{zCompositions::cmultRepl}})
is used to impute the inflated zeros. Otherwise, a small pseudo-count (\code{0.5})
will be added to all raw counts. Finally, either processed data will be converted
to relative abundance, then log of relative abundance will be used in the selection procedure.
The default value is \code{TRUE}}

\item{model}{Specifies which kind of survival regression model is built.
The options are ``\code{coxph}'' (\code{\link[survival:coxph]{survival::coxph}}) and
``\code{parametric}'' (\code{\link[survival:survreg]{survival::survreg}}). 
Make sure there is no zero survival time if choose ``\code{parametric}''.
The default value is ``\code{coxph}''}

\item{dist}{Specifies which kind of parametric distribution is used if
\code{model} is ``\code{parametric}''. The options include ``\code{weibull}'',
``\code{exponential}'', ``\code{gaussian}'', ``\code{logistic}'', ``\code{lognormal}''
and ``\code{loglogistic}''. For more details, please see
\code{\link[survival:survreg]{survival::survreg}}. The default value is ``\code{weibull}''}

\item{stopping_pvalue}{The threshold of p-value to stop the forward search.
The forward search where, at each step, a new taxon is added to the existing
balance, will be stopped if the resulting p-value is larger than \code{stopping_pvalue}.
The default value is \code{0.15}}

\item{sequential_test}{A logical variable indicating whether sequential testing, in addition to
\code{stopping_pvalue}, is used to stop the forward search. If \code{TRUE}, the forward search where, at each step,
a new taxon is added to the existing balance, will be stopped
if the new balance is not significantly different from the current one
in the survival regression model. The default value is \code{FALSE}}

\item{sequential_alpha}{The level of significance for the sequential testing
if \code{sequential_test} is \code{TRUE}. The default value is \code{0.25}}

\item{selection_criterion}{The criterion to select the global balance in the final model.
The options include ``\code{min_pvalue}'' and ``\code{min_decrement_pvalue}''. If ``\code{min_pvalue}'' is used,
the balance that has the smallest p-value along the forward selection path will be selected.
If ``\code{min_decrement_pvalue}'' is used, the decrement of p-value along the forward selection path will be calculated,
the balance before the first decrement that is smaller than \code{selection_threshold} will be selected.
The default value is ``\code{min_decrement_pvalue}''}

\item{selection_threshold}{A threshold of p-value decrement used for balance selection in the final model if \code{selection_criterion}
is ``\code{min_decrement_pvalue}''. The default value is \code{0.15}}

\item{quantile_plotted}{A vector specifies the quantiles of the final selected
balance that will be shown in the
survival plot. The default value is \code{c(0.25, 0.75)}}

\item{alpha}{The level of significance for survival plot of the final selected
balance. The corresponding confidence intervals will be shown in the survival
plot. The default value is \code{0.05}}
}
\value{
\code{balance_selection} return a list containing the following components:
\item{global_p}{Omnibus p-value from the global community-level association test, MiRKAT-S, which examines whether there is an overall shift in the microbiome composition (presence-absence status and abundance, encoded by Bray-Curtis and Jaccard distances) regarding the survival outcome}
\item{selection_path}{A matrix showing the forward selection path. If \code{sign}
is \code{1}, the corresponding taxon is in the numerator of the balance;
If \code{sign} is \code{0}, the corresponding taxon is in the denominator of the balance.
\code{p_value} is the p-value of the balance consisting of taxa up to the current taxon}
\item{survival_model}{An object of class \code{coxph} or \code{survreg} depending on the user's input, which is the final model with the final selected balance of microbiome}
\item{balance_name}{A list containing names of the taxa in the numerator and the denominator of the final selected balance}
\item{balance}{The value of the final selected balance of microbiome for each sample}
\item{survival_plot}{A plot of survival curves with confidence intervals, stratified by chosen quantiles of the final selected balance of microbiome}
}
\description{
\code{balances_selection()} identifies the optimal balance, i.e., the log-ratio of the geometric means of two sets of taxa that is most associated with the survival outcome using a greedy forward step-wise selection approach
}
\examples{

balance_selection(Surv_obj = gvhd, data = bacteria, mult_repl = TRUE,
model = "coxph", selection_criterion = "min_pvalue")

balance_selection(Surv_obj = gvhd, data = bacteria, mult_repl = FALSE,
model = "parametric", dist = "weibull", selection_criterion = "min_decrement_pvalue")
}
\references{
Rivera-Pinto, J., Egozcue, J. J., Pawlowsky-Glahn, V., Paredes, R., Noguera-Julian, M., & Calle, M. L. \emph{"Balances: a new perspective for microbiome analysis"}. MSystems 3.4 (2018): 10-1128

Fine, J. P. \emph{"Comparing nonnested Cox models"}. Biometrika 89.3 (2002): 635-648

Wahrendorf, J., Becher, H., & Brown, C. C. \emph{"Bootstrap comparison of non‐nested generalized linear models: applications in survival analysis and epidemiology"}. Journal of the Royal Statistical Society: Series C (Applied Statistics) 36.1 (1987): 72-81
}
