% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvhd.R
\docType{data}
\name{gvhd}
\alias{gvhd}
\title{Survival Data}
\format{
An object of class \code{Surv} with 63 rows and 2 columns.
}
\usage{
gvhd
}
\description{
The data is a survival object containing survival time and censoring/event information for 63 samples.
}
\keyword{datasets}
